🚧  #11 cookieにトークンを保存し、サーバーコンポーネントとクライアントコンポーネントを分離した
