+++ src/app/home/page.tsx
@@ -1,19 +1,14 @@
-'use client';
-
 import React, { useEffect, useState } from "react";
 import { useGetUser } from "../../hooks/useGetUser";
+import { cookies } from "next/headers";
 
 const UserHome = () => {
-  const [token, setToken] = useState<string | null>(null);
-  
-  useEffect(() => {
-  // ローカルストレージからjwt_tokenを取得
-  const access_token = localStorage.getItem("access_token");
-  setToken(access_token);
-  }, [])
+
+  // クッキーからjwt_tokenを取得
+  const access_token = cookies().get("access_token");
 
   // ユーザー情報を取得するカスタムフックを使う
-  const { data, error, isLoading } = useGetUser(token);
+  const { data, error, isLoading } = useGetUser(access_token);
 
   if (error) return <div>Error: {error.message}</div>;
   if (isLoading) return <div>loading...</div>;
