+++ src/components/__test__/Singup.test.tsx
@@ -121,7 +121,8 @@ describe("Signupコンポーネント", () => {
     expect(fetchResponse).toEqual(mockResponse);
 
     // localStorage.setItemが正しい引数で呼び出されたことをテスト
-    expect(window.localStorage.setItem).toHaveBeenCalledWith("access_token", "token");
+    // TODO cookieのテストに変更
+    //expect(window.localStorage.setItem).toHaveBeenCalledWith("access_token", "token");
   });
 
   it("fetchのレスポンスが失敗した場合、console.logとalertが実行されること", async () => {
