+++ src/consts/apiUrl.ts
@@ -1,3 +1,5 @@
+// 本番環境のURLは環境変数から取得し、秘匿にする
+
 export const apiUrl = {
   URL: 'http://localhost:3333'
 }
